backend/requirements.txt
requests==2.32.5
qdrant-client==1.16.2
sentence-transformers==5.2.0
PyMuPDF==1.26.7
python-dotenv==1.2.1
